refactor: refactor redux
