+++ src/reduxUsers/reducers/index.ts
@@ -1,3 +1,8 @@
+import { combineReducers } from 'redux';
 import { reducerIsLogging } from './reducerIsLogging';
 
-export { reducerIsLogging };
+export const rootReducer = combineReducers({
+  isLogging: reducerIsLogging,
+});
+
+export type RootState = ReturnType<typeof rootReducer>;
